+++ CHANGELOG.md
@@ -9,9 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Changeover to MIT license ([#2](https://github.com/scm-manager/scm-cas-plugin/pull/2))
 
+## 2.0.0-rc5 - 2020-03-19
+### Fixed
+- Redirection with anonymous access enabled fixed
+
+## 2.0.0-rc4 - 2020-03-19
+### Fixed
+- Hotfix for rc2: Redirection with anonymous access enabled fixed
+
 ## 2.0.0-rc3 - 2020-03-13
 ### Added
 - Add swagger rest annotations to generate openAPI specs for the scm-openapi-plugin. ([#1](https://github.com/scm-manager/scm-cas-plugin/pull/1))
 
 ## 2.0.0-rc2 - 2020-01-31
+### Fixed
 - Fix NPE on restart